requirements.txt
soco>=0.30.0
Pillow>=10.0.0
RPi.GPIO>=0.7.1
smbus2>=0.4.3
